ME-659: Moved AfterResult to the continue_strategies module.
