+++ clare/clare/retry/stop_strategies.py
@@ -3,26 +3,7 @@
 import abc
 import time
 
-
-class IContinueStrategy(object):
-
-    __metaclass__ = abc.ABCMeta
-
-    @abc.abstractmethod
-    def should_continue(self, attempt):
-
-        """
-        Parameters
-        ----------
-        attempt : Attempt
-
-        Returns
-        -------
-        bool
-            True if the client should continue and False otherwise.
-        """
-
-        pass
+from . import continue_strategies
 
 
 class IStopStrategy(object):
@@ -46,7 +27,7 @@ class IStopStrategy(object):
         pass
 
 
-class After(IContinueStrategy, IStopStrategy):
+class After(continue_strategies.IContinueStrategy, IStopStrategy):
 
     __metaclass__ = abc.ABCMeta
 
@@ -108,29 +89,6 @@ class AfterNever(After):
 
     def should_stop(self, attempt):
         return False
-
-
-class AfterResult(IContinueStrategy):
-
-    def __init__(self, predicate):
-
-        """
-        Parameters
-        ----------
-        predicate : callable
-            The callable must accept one argument and should return a
-            boolean.
-        """
-
-        self._predicate = predicate
-
-    def should_continue(self, attempt):
-        result = self._predicate(attempt.result)
-        return result
-
-    def __repr__(self):
-        repr_ = '{}(predicate={})'
-        return repr_.format(self.__class__.__name__, self._predicate)
 
 
 class AfterSuccess(After):
