+++ clare/clare/retry/tests/test_stop_strategies.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from nose.tools import assert_false, assert_true, raises
+from nose.tools import assert_false, assert_true
 
 from .. import stop_strategies
 from ..attempt import Attempt
@@ -96,37 +96,6 @@ def test_after_never_should_not_stop():
                       exception=None,
                       first_attempt_start_time=None)
     assert_false(stop_strategy.should_stop(attempt=attempt))
-
-
-def test_after_result_should_continue():
-    stop_strategy = stop_strategies.AfterResult(predicate=lambda x: x == 'foo')
-    attempt = Attempt(number=None,
-                      was_successful=None,
-                      result='foo',
-                      exception=None,
-                      first_attempt_start_time=None)
-    assert_true(stop_strategy.should_continue(attempt=attempt))
-
-
-def test_after_result_should_not_continue():
-    stop_strategy = stop_strategies.AfterResult(predicate=lambda x: x == 'foo')
-    attempt = Attempt(number=None,
-                      was_successful=None,
-                      result='bar',
-                      exception=None,
-                      first_attempt_start_time=None)
-    assert_false(stop_strategy.should_continue(attempt=attempt))
-
-
-@raises(TypeError)
-def test_after_result_predicate_accepts_one_argument():
-    stop_strategy = stop_strategies.AfterResult(predicate=lambda: None)
-    attempt = Attempt(number=None,
-                      was_successful=None,
-                      result=None,
-                      exception=None,
-                      first_attempt_start_time=None)
-    stop_strategy.should_continue(attempt=attempt)
 
 
 def test_after_success_should_stop():
